// helpers/types.ts
export enum eContractid {
    Example = 'Example',
    Unitroller = 'Unitroller',
    Comptroller = 'Comptroller',
    JumpRateModel = 'JumpRateModel',
    PErc20Delegate = 'PErc20Delegate',
    PERC20Delegator = 'PERC20Delegator',
    CEther = 'CEther',
    SimplePriceOracle = 'SimplePriceOracle',
    SimplePriceOracleChainlink = 'SimplePriceOracleChainlink',
    NewAssetsToken = 'NewAssetsToken',
    PubMiningRateModel = 'PubMiningRateModel',
    MarginSwapPool = 'MarginSwapPool',
    PriceOracleAggregator = 'PriceOracleAggregator',
    TestPriceOracle = 'TestPriceOracle', //moac
    // Config = 'Config',
    Controller = 'Controller',
    MockTXAggregator = 'MockTXAggregator',
    MSPStorage = 'MSPStorage',
    Capital='Capital',
    Liquidation='Liquidation',
    PUB='PUB',
    AToken = 'AToken',
}

export interface ICommonConfiguration {
    MarketId: string;
    ChainlinkAggregator: iParamsPerNetwork<ITokenAddress>;
}

export interface SymbolMap<T> {
    [symbol: string]: T;
}

export interface IRateConfiguration {
    MarketId: string;
    RateConfig: iParamsPerNetwork<ITokenAddress>;
}

export interface IRateModelPara {
    baseRatePerYear: string
    multiplierPerYear: string
    jumpMultiplierPerYear: string
    kink: string
}

export interface IReserveCollateralParams {
    baseLTVAsCollateral: string; //0.75
    liquidationThreshold: string; //0.5
    liquidationBonus: string; //1.08
}

export interface ITokenAddress {
    [token: string]: tEthereumAddress;
}

export interface IInterestRateStrategyParams {
    name: string;
    optimalUtilizationRate: string;
    baseVariableBorrowRate: string;
    variableRateSlope1: string;
    variableRateSlope2: string;
    stableRateSlope1: string;
    stableRateSlope2: string;
}

export interface IReserveBorrowParams {
    // optimalUtilizationRate: string;
    // baseVariableBorrowRate: string;
    // variableRateSlope1: string;
    // variableRateSlope2: string;
    // stableRateSlope1: string;
    // stableRateSlope2: string;
    borrowingEnabled: boolean;
    stableBorrowRateEnabled: boolean;
    reserveDecimals: string;
}
export interface IUnderyingInfo {
    initialAmount: string;
    tokenName: string;
    symbol: string;
    decimalUnits: number;
}

export interface IReserveParams extends IReserveBorrowParams, IReserveCollateralParams {
    //所依赖的资产
    underlying: IUnderyingInfo

    //cToken自身
    aTokenImpl: eContractid;
    reserveFactor: string;
    rateModel: IRateModelPara;
    initialExchangeRateMantissa: string;
    cTokenName: string;
    symbol: string;
    decimals: number;
    becomeImplementationData: string;
}

export enum SupportTokens {
    ETHER = 'ETHER',
    DAI = 'DAI',
    AAVE = 'AAVE',
    TUSD = 'TUSD',
    BAT = 'BAT',
    WETH = 'WETH',
    USDC = 'USDC',
    USDT = 'USDT',
    SUSD = 'SUSD',
    ZRX = 'ZRX',
    MKR = 'MKR',
    WBTC = 'WBTC',
    LINK = 'LINK',
    KNC = 'KNC',
    MANA = 'MANA',
    REN = 'REN',
    SNX = 'SNX',
    BUSD = 'BUSD',
    USD = 'USD',
    YFI = 'YFI',
    UNI = 'UNI',
    ENJ = 'ENJ'
}